Set the timezone when the `DateTime` object is created
Ticket: https://central.tri.be/issues/98654

If the Timezone is provided and the `$date` is not a UNIX Timestamp
format the timezone object should be set on initial value instead as if
not UTC is used as default.

For UNIX timestamps it can't be provided on Object creation only once
object is already created.