feat(Repository/ORM): Add action to filter the built ORM query.
This commit adds a dynamic "tribe_repository_{$filter_name}_query" action after the ORM query is
built and before it's returned. The action will fire whether the query is built by the ORM or by a a
query builder (e.g. a decorator). The action allows filtering functions to discriminate between the
two scenarios by passing the query builder flag and object, if any.

re https://central.tri.be/issues/125511