feat(ORM): Add `hash` method to get a hash of currently applied filters.
While the Repository/ORM will build a `WP_Query` object it will also set a number of filters and
actions on the query inner workings to speed up some queries. This means the `query_args`, both on a
`Repository` instance and on the `WP_Query` object it will produce, are not going to be a complete
representation of all filters the ORM is applying. The applied filters are stored in the `filters`
property on the ORM and that property, with exclusions and inclusions, is what the `hash` method
will use to build an hash.

re https://central.tri.be/issues/123950