fix(ORM): Add the query vars to the hash generation.
After the ORM filters have been applied an action is fired to allow integrations to apply further
filtering to the query. This fix adds a `WP_Query $query` parameter to the `hash` method to allow
generating an hash that includes a built query query vars in its generation. The `$settings`,
exclusions and inclusions, will apply to the query variables too.

re https://central.tri.be/issues/123950