fix(ORM): Make hash generation more flexible and robust.
This fix decouples the gathering of the hash data, in the `Tribe__Repository::get_hash_data` method,
and its use, in the `Tribe__Repository::hash` method. It also makes the hash generation more robust
by running a recursive key sort on the generated array to ensure consistency across queries  that
might be the same in substance but differ in order.

re https://central.tri.be/issues/123950