feat(Tribe__Log): Allow filtering out wp-cli logging.
Previously we would be logging to wp-cli if available. Logging errors would `exit` the PHP thread
and this would be disruptive in the context of tests. This fix adds the `tribe_common_log_to_wpcli`
filter to allow controlling that behaviour and deactivates it in the context of the tests.