feat(Repository/ORM): Avoid query rebuilds.
This improvment makes it so that the query, an instance of the `WP_Query` class, built by the
repository will only be rebuilt if required by a change in the arguments and filters applied to the
repository instance. When changing from ID based fetches to post objects based fetches the
repository will try to re-use as much as possible without running a new query.