feat(Context): Add view-related locations
This commit adds some View-related locations to lay the base for the View Refresh work. Beside some
use of closures to populate some locations I've added `view` and `view_data` to manage the state of
Views; `view` is a slug like `list` while `view_data` is an array that will contain any information
regarding a View. The purpose of `view_data` is to avoid having to endlessly add locations to the
context to support one more view. In this first iteration the accessor of the view data is not
implemented as I've laid out the bare-bones requirement to make tests pass.