Restore Composer autoloader On version 4.9.6 composer loader was generated using `composer --no-dev -a` which introduced the changes we are reverting here.
That seems to have introduced bugs for customers when it was trying to activate the `Composer\Autoload\ClassLoader::setClassMapAuthoritative()` as the method seems to be non-existent.

It doesnt happen on any of our local testing, QA or Smoketesting. More digging needs to happen here to allow us to figure out what makes this happen.

@sc0ttkclark thinks it has to do with problems with other Composer autoloader conflicting.`