refactor(Main::post_id_helper): refactor the code to remove redundany if/else, clean-up tests
This commit splits some if/else flows that were really not applying due to early returns and
cleans-up the test code a bit.Worth noting is the removal of the post creation in the data provider
of the `Post_Id_Helper_Test` class that would cause the created post object to be left-over after
the  test case ran. Data providers run **before** the test methods and **before** the test cases
`setUp` methods and any side effects produced by the data providers will not be torn down.