fix(Dependency): remove version number when redirecting users to latest version
While trying to be helpful, in the context of plugin below the parent plugin minimum required
version (e.g. PRO is not up to the minimum required version by TEC), we show a message linking to
the latest version download and provide a version number. That version number is, in the case above,
the minimum required version, not the latest version. This fix removes the version to avoid
providing a confusing message in this case.

re https://central.tri.be/issues/126615