refactor(Tribe__Context): add more locations to the context
This commit adds more locations to the Context and moves the definitions of those locations to a
separate file. Locations, even the static ones, need to be initialized at build time to cope with
values that cannot be inlined in a `static` declaration (e.g. a closure). Moving them to a separate
file makes sure we keep the `Tribe__Context` class file "lean", avoid loading the locations at
auto-load time and get a small optimization bonus.

re https://central.tri.be/issues/126071