fix(Tribe__Repository.php): fix hashing issue preventing queries from correctly re-running
This fix adds the tests and code to make sure queries will correctly re-run when the filtering
parameters or arguments change. Not only that: queries should re-run for the `found` method too when
either the arguments are changed or when the found rows request did not run in the first place. The
contrary applies as well: the request run from the `found` method should not "obfuscate" the request
that should run for the following methods.

re https://central.tri.be/issues/127710