fix(Tribe__Rewrite): remove eager call to `setup`
In B19.07 I've added an eager call to the `setup` method in the `instance` (singleton
implementation) of the class. That method is called in many places in the code and before plugins in
need of filtering the rewrite bases (e.g. PRO) have even hooked or initialized. The `setup` method
should be called lazily, and on-demand, when the setup is required.