feat(Cost_Utils.php): Add the `parse_separators` method
The `Tribe__Cost__Utils::parse_separators` method comes in handy when trying to parse cost values
from cost strings that might have been generated in the context of another locale. With some basic
logic we might infer that `23,89` is using `,` as a decimal separator even in an `en_EN` locale. On
the same note, while in the context of a `de_DE` locale we might infer `23.89` is using `.` as a
decimal separator. The method, due to inherent limits to this guessing, will not be able to come to
a ny definitive conclusion, and will fall back on the separators defined by the current locale, when
working on non numeric vvalues (e.g. `Free`) or "ambiguous" values (e.g. `23,899`) where the 3
numbers after the separator do not allow to clearly distinguish it.

re https://central.tri.be/issues/98061