fix(Tribe__Cache.php) store last occurrence microtime
So far we've stored the last occurrence time using an integer. This is
fine in "user" times but can lead to missing cache invalidation in the
context of rapidly executing code that could cache, invalidate and read
in the context of the same second.
Storing the last occurrence as microtime avoids this issue.