feat(tooltip assets): Don't use the same name for two assets.
A few things happening here. First, we cahnged the style and script names so they are different to
avoid any checks incorrectly tossing one out. Then we instituted a check for assets taht had already
been registered - no need to do so twice. We also ustilized `wp_style_is` and `wp_script_is` for
more accurate information on regeistered assets. Then we went into the script and tweaked a few
things - mainly removed a `.` from the beginning of the class name - since that isn't a valid class
name and won't match the className check later in the script.

:ticket: https://central.tri.be/issues/130658