Remove connector from the normal Promoter provider
Ticket: https://central.tri.be/issues/134862

Remove the call to `determine_current_user` in the Promoter provider as
tht one is executed once all the plugins has been loaded with priority
of 1, when some plugins might use authentication functions on priority 0
or before, we need to set this filter ASAP on the plugin load so when
the authentication is called again the filter can be fired.

Move the callback for `tribe_promoter_secret_key` so is executed when
the secret key is not defined.