fix(Rewrite) use the base to avoid conflicts btwn bases
Ticket: https://moderntribe.atlassian.net/browse/TEC-3254

This update build the localized matchers, those that can change
depending on the site language or the user settings, to make sure
that bases that would produce the same slugs will not conflict.

E.g. in Swedish (I'm told) the singular and plural version of "events"
is the smae and this creates issues with the management of pretty,
canonical, URLs and their manipulation.

Furhtermore The Events Calendar allows users to set the single and
plural event slugs to the same value, e.g. `course`, and this too
would cause issues.

We use the base, e.g. `archive` or `single` to discriminate the slugs
during lookup and then conflate them when we match them in canonical
URL building.