Use int for first arg of the Monolog ErrorHandler
Our instantiation of the `Monolog\Handler\ErrorHandler` is incorrect as per the docblock in both versions 1.x and 2.x of Monolog. The first argument should be an `int` and we were passing `null`. With Monolog 2.x, the type hinting enforces the use of an `int` where we should have been using one all along.

See: [TEC-3638]