Fix non_persistent_keys checks so they always reference the correct cache key
This is a simple PR in it's intent: The non-persistent keys aren't getting set to the non-md5'd key properly in the set() method.

The isset() logic for non-persistent keys has also been updated to reduce overhead and complexity.

This PR includes a new test that confirms this fix and provides checks for various values/types to confirm they work as *currently* expected.