feat(Cache) add support for chunkable transients
This commit adds the code, and tests, to support "chunkable" transient.

A "chunkable" transient is one where we know that the cost of building
the transient value will, 80% of the times, be more than the one
required to fetch each transient row from the db.

The function will be completely skipped when object caching is active
and developers will have to make a choice to use the API with knowledge
of the above.