Make sure cache is retrived correctly
Make sure the pieces from the cache start at `0` as all the index start
on that number we need to make sure we get the first element in this
case the element with index `zero` from there we build the subsequent
pieces.

Then the function if was able to build the elements out of the cache it
should return the values from the cache instead of returning `true`.

As the caller to the `chunkable_transient` should build the results out
of the cache.