Defer to WP domain before checking $_SERVER
This resolves an issue on sites hosted on servers whose `$_SERVER['SERVER_NAME']` differs from the domain specified in `get_option( 'siteurl' )`. For the purposes of license validation, we only care about the siteurl and should only default to the `$_SERVER['SERVER_NAME']` variable if siteurl is filtered such that it is empty.

:ticket: [TCMN-122]
:movie_camera: https://d.pr/v/s1NToE