feat(Assets) print localization data when printing group
The `Assets::print_group` method would correctly print scripts when
forcing the print of a registered script, but would not print the
localization data the script might require along with it.

This commit adds the localization data printing, as an opt-in option by
means of a boolean flag, to ensure back-compat is maintained and new
code will be able to force the printing of the localized data without
issues.