fix(Main.php) execute after plugins_loaded
When running Common code in the context of a plugin activation, then
Common code would load on the activation hook.
Common relies on the `plugins_loaded` hook to correctly load its code,
and that would not fire as, in the context of a plugin activation, the
`plugins_loaded` action would fire **before** the activation one.

This commit adds a fallback to make it so that Common will load
correctly in the context of a plugin activation by firing the
`Tribe__Main::plugins_loaded` method if `did_action( 'plugins_loded'
);`.