feat(Utils/Array) add the usearch method
Add the `Array::usearch` method to the class to fill the gap left by the
non existence of an `array_usearch` function that accepts a callback and
will bail on first match. The work-around of either a `foreach` or an
`array_filter` followed by a `reset` are not satisfactory or performant.