Allow for the suppression of PUE notices for specific plugins
The additions of these three filters will allow for the suppression of license notices where appropriate. Example use case: tribe-common currently triggers license notices for Restrict Content Pro when TEC or ET is active, and that is not the desired path for notification for RCP at this time. These filters will allow RCP to extract itself from the list of plugins that trigger that notice.