fix(Rewrite) laxer rewrite slugs matching
Handle the fact that rewrite slugs might contain the localize version or
not updating the code to have a more flexible resolution of canonical
URLs when there are, or there are not, slug translations available.

This would be evident, and cause issues, when the translation slug is
generated with 2 or more languages, but the resolution happens with 1
language only.

The fix leverages the fact that, by default, the English US slug will
always be part of the rewrite rules.