fix(Utils/Taxonomy) avoid fatal on invalid terms
This fixes an issue where the function would assume all the terms
returned from the `get_terms` function would be valid `WP_Term`
instances; due to the filtering that happens in the `get_terms`
filter, that might not be true.
The fix consists in dropping invalid term results from the prime
operation.