Enable auto-updates for premium plugins
This change (from [here](https://github.com/stellarwp/uplink/pull/57)) is super simple. We're already interacting with the update transient to show that there are updates pending. To enable the `Enable Auto-updates` link, we just need to add an entry when there are _no_ updates as well.  Dead simple and we didn't even realize that.