refactor(Controller_Test_Case)
Rename the `setUp` method to `setUpTestCase` and annotate it with `@before`.
Rename the `tearDown` method to `tearDownTestCase` and annotate it with `@after`.
Not overriding the base test case `setUp` and `tearDown` methods allows the `Controller_Test_Case` to be use across different PHPUnit versions.
Newer versions would require the `: void` return type that older versions will not support.