fix(src/modules) day-picker value
This updates the element to not rely solely on the `selectedDate` from
the `useState` hook.
The date picker will render a number of times with an empty `value`
property that will, correctly, default the date to today.
That initial missing `value` will dictate the `selectedDate` for the
component making the use of `setSelectedDate` the only way to update its
value.
This changes the date picker to hydrate correctly from different,
folllowing, `value` properties and to handle the PHP and JS date-picker
format correctly.