build(package.json) update file-loader
The `file-loader` dependency, in its previous version `^1.1.6`, would
require `webpack-cli` v4. The `@wordpress/scripts` dependency would,
instead, require `webpack-cli` v5. When encountering such different
requirements `npm` would use peer-dependencies for the less requested
(WebPack v5 in the context of Common) and root dependencies for the more
compatible or requested dependency (WebPack CLI v4 in the context of
Common). This would make it so that the command to run `wp-scripts
build` would use the **root** `webpack-cli` (v4) and fail.

The update to the `file-loader` dependency reconciles the issue with no
change in functionality, but allowing Common to build correctly with its
own dependencies, not using the ones coming (possibly) from the context
(e.g. the TEC plugin that contains it).