test(classy_jest) fix interaction in tests
Some code cleanup and fixing of the interaction with the popover to
dismiss it.

Dispatching the `keyDown` Escape event on the popover is the correct
approach (matching the `closeOnEscapeRef` behaviour in WordPress code),
using the `@testing-library/react` `fireEvent` API would not work, the
reason seems to be the `fireEvent` would not use the deprecated
`keyCode` property of the Event used by WordPress to check for the
pressed key.