test(classy_jest) wip - timezone selection
The testing of this component poses two challenges:
1. Selecting from it the new option
2. Refactoring the component to pick the timezone choices from the store

The second in particular requires wrapping the component in a registry
provider for the purpose of testing it.
Both challenges are still open but should be close to what is required
to test the components in relative isolation and with
as-close-to-real-use as possible approach.