test(classy_jest) wrap interaction testing
This wraps the discovery of how to correctly handle interactions like
the ones based on popovers and how to fix the class names generated
by React in rendered components.

When it comes to the popover part, the discovery was the the
`@testing-library/user-event` package will implicitly call `act()` when
using interactions like `user.click`, but will not wait for the
component complete transitions to complete.
In the case of the popover, the transition from closed and not showing
to open and showing is represented by the component having the
`is-positioned` class assigned.
The use of `waitFor`, again from the `@testing-library/react` package,
allows waiting until an expectation (an `expect` function call) does not
throw an error. In this case the function is used to wait for the
popover component to be found with the `is-positioned` class assigned.

On the stable snapshot front, the WordPress components use the
`@emotion/react` and `@emotion/styled` libraries that will generate, on
each render, unique-by-design classes on each component. The classes
would be different on each render, making snapshot testing moot. The
remedy for this is the `@emotion/jest` package that will fix the
generated classes to always-equal values across test runs.