feat(Classy) start work on Core stores registration
When not in Block Editor context, then the Core WP stores will not be
available. The components used in the Classy applicaiton, both coming
from WordPress and custom-defined, rely on `select` and `dispatch`
functtions from the Core stores (e.g. the
`select('core/editor').getEditedPostAttribute('type')` code). This
updates the registry creation logic to either get (in Block Editor
context) or create (outside of the Block Editor context) the Core
registry and required stores and then register the Classy registry as a
sub-registry to it.
This simplifies the registry structure by keeping the same system in
place.

The current solution is not complete, though, as the initialization
phase normally handled by the `ExperimentalEditorProvider` component
from WordPress core is still not taken care of.
This commit is a good-enough start of the exploration and will provide a
start for the next iteration to find out how to create the Classy
provider correctly.