refactor(Classy) better default registry use
The changes remove the hack I had previously used to get hold of the
WordPress default registry (the one used by WordPress to handle all its
stores in the Block Editor context) to align with the default WordPress
API based around the use of the `withRegistry` HOC.

The refactor creates a Classy provider: a combination of a registry
consumer (the `registry` instance provided by the `withRegistry` HOC)
and a `RegistryProvider` (again from WordPress core) to pass-down, and
capture, the current default registry.

The change was prompted by the idea to simplify the code and leveraging
the fact that the `withRegistry` HOC will get, and if not already
present, build, the default registry to provide it to the wrapped
components. For the time being this removes the need to curate and
set up a dedicated registry for the Classy application both in and out
of the Block Editor context.

Functionally, the components and fields already part of the application
are not affected.

This makes possible integration testing of components, in Common and in
other plugins, that implicitly (through the `useRegistry` hook) or
explicitly (through the `useSelect` and `useDispatch` hooks) rely on the
registry instance.