fix(Classy) hide BE welcome guide
When looking at a post type supported by Classy, then the Block Editor
welcome guide should be hidden.

Hiding the Welcome Guide on the frontend either using CSS or JS would
"flash" it.
This code updates the user meta controlling the persisted user
preferences without changing them, making sure the welcome guide will
not show if the user is editing a post type managed by Classy and has
never dismissed the welcome guide before.

A catch: this is the best solution I could come up with that would
cleanly do this, but a user that is editing, as their first post, a post
managed by classy will have the preference saved in **local storage** of
the browser, and will not see the Welcome Guide on other posts.
If the user has already seen the welcome guide, then they will not be
affected.

A user that wishes to see the welcome guide again can just do so from
the standard Block Editor menu.