fix(Classy) tz dropdown issue
This fixes an issue where the timezone dropdown in Classy context would
not show correctly when using the `UTC+0` timezone. The fix consists on
using the the `Timezones` utils class and relying on that to always
provide a correct timezone string. The issue was caused by WordPress
storing the timezone in the `timezone_string` option, and UTC offsets in
teh `gmt_offset` option.