feat(Classy) add text presentational class
Add a `text` class to allow styling inputs that are replaced with text
entries correctly. This will happen when either the text is part of the
control and represents a value that cannot be changed, or when a select
control is replaced with a text when only one option will show.