fix(Classy) date/time start/end issues
This updates the logic used in the `StartSelector` and `EndSelector`
components to handle and pass around more precise information about the
user interaction with the start / end date / time selectors.

This allows fields using the components (e.g. the `EventDateTime` one
from TEC) to more precisely model and consume what field was directly
updated and what field was updated indirectly, as a consequence of
anoter update.