feat(KV Cache) first iteration
This objective of this first iteration of the key-value cache API is to
provide a transparent API to store computationally intensive and
possible large string values across requests.

When object-caching it's available the API will use it, else it will use
a custom table with a pretty simple structure.

The unique API is the `tec_kv_cache()` function, see more in the
documentation.

Following a WP VIP lead, I've set the minimum accepted expiration time
to 300 seconds, 5 minutes, else this is not the cache to use.

Furthermore the cache key column is set to a length of 191 chars
following the example of the `wp_postmeta.meta_key` column: 255 chars
long, but only 191 are indexed. I've picked the lower value to play on
the safe side.