This PR changes the delete logic to rely on wp_delete_post exclusively.
The logic now just modifies the `force` paramater to determine if the entity should be permanently deleted or not.
All of teh other logic is still in place, including the filter for converting trashed to permanent.

Note: `wp_delete_post()` only auto-converts trash to delete for "post" and "page" post types.