test(classy_jest) better @wordpress/data mocking
Following [a comment found in the Gutenberg
repository](https://github.com/WordPress/gutenberg/issues/42442), update
the testing setup to reduce the number of mocks required in component
testing to a partial mocking of the `@wordpress/data` package. As a
first application update the CurrencySelector tests simplifying the
structure of the test.