fix(Classy) use classy if no post
This fixes an issue where the Classy editor would not be used to show
Occurrences part of an Event. In that case, the Occurences post ID would
not resolve to a post since and the ECP-controlled methods for getting
the Event post ID from the Occurence provisional post ID are not yet
available.