feat(Classy) restore visual toggle
This restore the "Visual" toggle functionality that allows hiding the
Classy element to show the Block Editor and viceversa.
In this version, the toggle will not take care of setting up or
preparing the block editor blocks in any way, the block editor will show
empty of blocks.
The Visual toggle button will only show if the user adds
`&classy_enable_visual=1` to the URL.