fix: reset translation state correctly
This fixes the logic that would load plugin translation files from
custom paths part of the plugins to work correctly.
Starting from WordPress 6.7.1, **not** 6.7.0 which is the currently
minimum supported version, the `load_plugin_textdomain()` function will
reset the global `l10n`; the first fix consists in making sure that
reset is replicated in the context of sites running WordPress 6.7.0.

The second fix is needed regardless of the WordPress version: when a
translation is required early (e.g. a call to the `__()` function for
the plugin domain) then the translations are loaded early and, **in the
registry** set to `false` to indicate no translation file was found for
that domain and locale.
The `load_plugin_textdomain()` function will not reset that value in the
registry, but it will set the registry to include a custom path to look
for the plugin file. That custom path will not be used, though, since
the translation registry will not look up the files again. The
`$wp_textdomain_registry->set()` call overrides that cached value
providing now a translation file for the domain and locale.