fix: IN queries with spaces support
This fixes an issue where the repository, using `WP_Query` under the
hood, would build the incorrect `IN` or `NOT IN` query SQL when using
`by`/`where` methods like `meta_in`/`meta_not_in` with values containing
spaces.

An exmaple is `->by( 'meta_in', 'First Name')` where the built SQL would
be `AND meta_value IN ('First', 'Name')` resulting in a non-match for
"First Name".

This is due to the fact that, in the `WP_Meta_Query` class the `IN` and
`NOT IN` queries will split the meta value ("First Name" in the example)
at commas and spaces resulting in two meta values: "First" and "Name".
The split guard is `! is_array( $meta_value )`, so the casting to array
in the repository solves the issue as explicit use of an array would:
```
->by( 'meta_in', [ 'First Name' ] )
```
The problem is that a huge portion of the code relies on the use of
`meta_in` and similar and passes values indifferently as
`string|string[]` leading to the issue when a `string` value contains a
space (the guard does not apply).